splitting up config, styling with gruvbox
