+++ lua/init.lua
@@ -1,0 +1,1 @@
+require 'remaps'
