+++ lua/sld/gruvbox.lua
@@ -1,0 +1,7 @@
+return {
+  'f4z3r/gruvbox-material.nvim',
+  name = 'gruvbox-material',
+  lazy = false,
+  priority = 1000,
+  opts = {},
+}
