+++ lua/sld/init.lua
@@ -1,0 +1,6 @@
+require 'sld.remaps'
+require 'sld.options'
+require 'sld.lazy'
+require 'sld.autocmd'
+require 'sld.plugins'
+require 'sld.style'
